Created a resources folder containing one type of every pic format
