--- src/test/resources/Foo
@@ -1,1 +1,0 @@
-asasdfasdfasdf
